animation-timing/test_framerequestcallback_in_onesecond.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>FrameRequestCallback will be called within 1 seconds at the latest</title>
<link rel="author" title="Kensaku Komatsu" href="mailto:kensaku.komatsu@gmail.com" />
<link rel="help" href="http://www.w3.org/TR/animation-timing/#framerequestcallback"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="stylesheet" href="/resources/testharness.css" />
<script>
  setup({explicit_done: true});

  // avoid script errors
  if(window.requestAnimationFrame === undefined) {
    done();
    return;
  }

  // set true when FrameRequestCallback called.
  var called = false;
  var fr_callback = function(){
    called = true;
  }

  requestAnimationFrame(fr_callback)
  

  // after 1 second, assert called is true or not
  setTimeout(function(e){
    var msg = "FrameRequestCallback has been called within 1 seconds.";

    test(function(){
      assert_true(called, msg);
    }, msg);
    done();
  }, 1000);
</script>
</head>
<body>
<h1>Description</h1>
<p>This test validates FrameRequestCallback will be called within 1 second at the latest.</p>

<p>If callback hasn't been called after 1 second, there will be a bug in hardware or software.</p>

<div id="log"></div>
</body>
</html>
